refactor(config): move config helpers into separate service
