+++ config/services.yaml
@@ -8,10 +8,10 @@ services:
     resource: '../src/*'
     exclude: '../src/{DependencyInjection,Interface,Model,Tool,PhpToTsBundle.php}'
 
+  Brainshaker95\PhpToTsBundle\Service\Configuration:
+    bind:
+      $config: '%php_to_ts%'
+
   Brainshaker95\PhpToTsBundle\Service\Filesystem:
     bind:
       $projectDir: '%kernel.project_dir%'
-
-  Brainshaker95\PhpToTsBundle\Service\Dumper:
-    bind:
-      $config: '%php_to_ts%'
