+++ src/Service/Dumper.php
@@ -3,11 +3,6 @@
 namespace Brainshaker95\PhpToTsBundle\Service;
 
 use Brainshaker95\PhpToTsBundle\Interface\Config;
-use Brainshaker95\PhpToTsBundle\Interface\FileNameStrategy;
-use Brainshaker95\PhpToTsBundle\Interface\SortStrategy;
-use Brainshaker95\PhpToTsBundle\Model\Config\FileType;
-use Brainshaker95\PhpToTsBundle\Model\Config\FullConfig;
-use Brainshaker95\PhpToTsBundle\Model\Config\Indent;
 use Brainshaker95\PhpToTsBundle\Model\TsInterface;
 use Brainshaker95\PhpToTsBundle\Tool\Str;
 use PhpParser\Error;
@@ -22,40 +17,19 @@ use Symfony\Contracts\Service\Attribute\Required;
 class Dumper
 {
     #[Required]
+    public Configuration $config;
+
+    #[Required]
     public Filesystem $filesystem;
 
     #[Required]
     public Visitor $visitor;
 
-    private FullConfig $config;
-
     private Parser $parser;
 
-    /**
-     * @param array{
-     *     input_dir: string,
-     *     output_dir: string,
-     *     file_type: FileType::TYPE_*,
-     *     indent: array{
-     *         style: Indent::STYLE_*,
-     *         count: int,
-     *     },
-     *     sort_strategies: class-string<SortStrategy>[],
-     *     file_name_strategy: class-string<FileNameStrategy>
-     * } $config
-     */
-    public function __construct(array $config)
+    public function __construct()
     {
         $this->parser = (new ParserFactory())->create(ParserFactory::ONLY_PHP7);
-
-        $this->config = new FullConfig(
-            inputDir: $config['input_dir'],
-            outputDir: $config['output_dir'],
-            fileType: $config['file_type'],
-            indent: new Indent($config['indent']['style'], $config['indent']['count']),
-            sortStrategies: $config['sort_strategies'],
-            fileNameStrategy: $config['file_name_strategy'],
-        );
     }
 
     /**
@@ -72,7 +46,7 @@ class Dumper
         Config|string|null $configOrDir = null,
         ?Config $config = null,
     ): void {
-        $config = $this->getConfig($configOrDir instanceof Config ? $configOrDir : $config);
+        $config = $this->config->merge($configOrDir instanceof Config ? $configOrDir : $config);
 
         $dir = $this->filesystem->makeAbsolute(
             is_string($configOrDir) ? $configOrDir : $config->getInputDir(),
@@ -121,7 +95,7 @@ class Dumper
             return;
         }
 
-        $config           = $this->getConfig($config);
+        $config           = $this->config->merge($config);
         $fileType         = $config->getFileType();
         $fileNameStrategy = $config->getFileNameStrategy();
         $pathPrefix       = $config->getOutputDir() . DIRECTORY_SEPARATOR;
@@ -168,25 +142,5 @@ class Dumper
         $traverser->traverse($statements);
 
         return $this->visitor->getTsInterfaces();
-    }
-
-    private function getConfig(?Config $config = null): FullConfig
-    {
-        if (!$config) {
-            return $this->config;
-        }
-
-        if ($config instanceof FullConfig) {
-            return $config;
-        }
-
-        return new FullConfig(
-            inputDir: $config->getInputDir() ?? $this->config->getInputDir(),
-            outputDir: $config->getOutputDir() ?? $this->config->getOutputDir(),
-            fileType: $config->getFileType() ?? $this->config->getFileType(),
-            indent: $config->getIndent() ?? $this->config->getIndent(),
-            sortStrategies: $config->getSortStrategies() ?? $this->config->getSortStrategies(),
-            fileNameStrategy: $config->getFileNameStrategy() ?? $this->config->getFileNameStrategy(),
-        );
     }
 }
